Question from student
Question from student was:

The way I understood it is that I should be able to run cross validation on the entire data set since each iteration fits the model on K-1 folds and tests the model on the Kth fold. I was completing the Boston housing regression mini project and noticed that when I use CV on the training data set as split by sklearn.model_selection.train_test_split I get very similar MSE values with / without CV

But if I use the entire dataset - I get much higher MSE values when using CV. I am not sure why that is the case. Here's the notebook.